+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+---------------------
+##### Fixed
+- If repeat has no template, duplicate and conflicting ordinal attributes are added.
+
 [4.12.0] - 2016-10-17
 ---------------------
 ##### Added